data: daily_prices.volume BIGINT -> DOUBLE (Polygon reports FRACTIONAL volume)
User catch: some TRF prints carry fractional share counts and the tables were
never adjusted for it.

⚠ WE WERE LOSING IT. Polygon's daily flat files began reporting fractional volume
on 2026-02-23 — a hard changeover, found by bisection:
    2026-02-18/19/20   0.0% fractional
    2026-02-23 onward  ~85% fractional
A BIGINT column floored every one. 2026-08-07 ticker A: the CSV says
1,525,223.372108; we stored 1,525,223.

DOUBLE end to end now (daily_prices -> daily_adjusted -> the already-DOUBLE
avgvol20/avgvol20_prior). Exact round-trip verified against the CSV:
A 1525223.372108, AA 5991598.536414, AAPL 34437191.007859. 1,215,019 fractional
rows — 86.63% of everything from 2026-02-23, 0% before.

Aggregate impact is small (604,718 shares across 49.4M rows) and no gate moves on
it — but it is the SAME DEFECT CLASS as split_adjusted_prices.adj_volume, where a
non-integer quantity in an integer column truncated to 0 and silently deleted 436
universe ticker-days. A raw share count is no longer an integer; the type should
say so.

daily_prices was DROPped and re-ingested from data/daily_aggregates/*.csv.gz (the
archive is the source of truth): 49,419,332 in, 49,419,332 out. All materialized
tables and mr_candidate_1s_v2 rebuilt; universe parity held EXACTLY at
1,431,802 / 7,585 / 2,513. Full acceptance suite passes.

⚠ GOTCHA, now documented in the view file: DuckDB caches a view's column types, so
both daily_episodes views had to be re-created after the base-table type change —
and the failure surfaces in the CONSUMER ("Contents of view were altered: types
don't match!" from build_mr_candidate_1s.fsx), not in the view itself.

NB the 1s slim corpus stores vwap/volume as FLOAT (single) and ALREADY carried
fractional volume — 6.27% of bars, down to 0.000994 — so the intraday path never
had this bug; the engine widens to Double on read. Left as FLOAT deliberately:
~7 significant digits is ~1e-7 relative error against gates comparing at ~1e-4,
and DOUBLE would add ~40GB to a 190GB corpus.

+++ TradingEdge.Database/Database.fs
@@ -16,7 +16,7 @@ type DailyPriceRow = {
     high: float
     low: float
     close: float
-    volume: int64
+    volume: float
     transactions: int64
 }
 
@@ -204,7 +204,7 @@ let ingestDailyPricesFromCsvGz (connection: IDbConnection) (filePath: string) : 
         FROM read_csv('{filePath}',
             columns = {{
                 'ticker': 'VARCHAR',
-                'volume': 'BIGINT',
+                'volume': 'DOUBLE',
                 'open': 'DOUBLE',
                 'close': 'DOUBLE',
                 'high': 'DOUBLE',
@@ -235,7 +235,7 @@ let ingestDailyPricesFromGlob (connection: IDbConnection) (globPattern: string) 
         FROM read_csv('{globPattern}',
             columns = {{
                 'ticker': 'VARCHAR',
-                'volume': 'BIGINT',
+                'volume': 'DOUBLE',
                 'open': 'DOUBLE',
                 'close': 'DOUBLE',
                 'high': 'DOUBLE',
